Use 1 CPU and 8 GB of memory
Now that we are using less memory in the app I reduced the memory
to 8 GB. This allowed me to lower the number of CPUs requested
as well. Since the open source shiny server only uses 1 CPU.
Previously we allocated 4 CPU to get to the higher memory
threshold previously required by the app.